requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.25.0
pymongo==4.6.0
python-dotenv==1.0.0
pydantic==2.8.0